validate that async calls work across wasm boundary

+++ emulator/chip-8/src/wasm.zig
@@ -1,2 +1,2 @@
-pub extern fn consoleLog(message: [*]const u8, length: c_uint) void;
+pub extern fn consoleLog(message: [*]const u8, length: usize) void;
 pub extern fn getRandomSeed() c_uint;
